Add observed attributes and DTS file
