+++ addons/van_component/index.html
@@ -7,29 +7,39 @@
 </head>
 <body>
   <h1>Webcomponents + VanJS = ❤️</h1>
-  <counter-add></counter-add>
-  <counter-row>✅</counter-row>
+
+  <p>Here is a demo component:</p>
+  <demo-component></demo-component>
+  <div>
+    <p>And here is the standalone component it uses:</p>
+    <font-preview size="20" color="orange"></font-preview>
+  </div>
 
   <script type="module">
-    import van from "vanjs-core";
-    import {createComponent} from "./src/van-component";
+    import van from "vanjs-core"
+    import {createComponent} from "./src/van-component"
 
-    const {button, div, slot} = van.tags;
+    const {input, option, select, span} = van.tags
 
-    const Counter = ({plus, minus}) => {
-      const counter = van.state(0)
-      const dom = div(
-        slot(),
-        " ", counter, " ",
-        button({onclick: () => ++counter.val}, plus ?? "+"),
-        button({onclick: () => --counter.val}, minus ?? "-"),
-        button({onclick: () => dom.remove()}, "❌")
+    createComponent("font-preview", ({size, color}) => span({style: () => `font-size: ${size.val}px; color: ${color.val};`}, " Hello 🍦VanJS")
+      , ["size", "color"]);
+
+    createComponent("demo-component", () => {
+      const size = van.state(16), color = van.state("black")
+      return span(
+        "Size: ",
+        input({
+          type: "range", min: 10, max: 36, value: size,
+          oninput: e => size.val = e.target.value
+        }),
+        " Color: ",
+        select({oninput: e => color.val = e.target.value, value: color},
+          ["black", "blue", "green", "red", "brown"]
+            .map(c => option({value: c}, c)),
+        ),
+        van.tags["font-preview"]({size: size, color: color})
       )
-      return dom;
-    }
-
-    createComponent("counter-row", Counter);
-    createComponent("counter-add", () => button({onclick: () => van.add(document.body, van.tags["counter-row"]("❤️"))}, "Add"));
+    });
   </script>
 </body>
 </html>
