+++ addons/van_component/src/van-component.d.ts
@@ -1,0 +1,7 @@
+import { State } from "vanjs-core";
+
+export declare const createComponent: (
+  name: string,
+  component: (attributes: Record<string, State<string>>) => HTMLElement,
+  observed?: string[]
+) => void;
